Refactor table listing for key results
