+++ assets/js/pages/ObjectivePage/KeyResults.tsx
@@ -3,6 +3,8 @@ import React from 'react';
 import { useTranslation } from 'react-i18next';
 import { useQuery, gql } from '@apollo/client';
 
+import Table from './Table';
+
 const GET_KEY_RESULTS = gql`
   query GetKeyResults($objectiveID: ID!) {
     keyResults(objectiveID: $objectiveID) {
@@ -15,45 +17,6 @@ const GET_KEY_RESULTS = gql`
 interface KeyResult {
   id: string;
   name: string;
-}
-
-function KeyResult({keyResult} : {keyResult: KeyResult}) : JSX.Element {
-  return (
-    <div className="mt-4 flex gap-2 items-center">
-      <div>
-        <div className="font-bold">{keyResult.name}</div>
-      </div>
-    </div>
-  );
-}
-
-interface Header {
-  id: string;
-  label: string;
-}
-
-interface Row {
-  [key: string]: string;
-}
-
-function Table({headers, columnClasses, rows}) : JSX.Element {
-  return (
-    <div className="mt-4 flex flex-col gap-2 items-center rounded border border-stone-200">
-      <div className="w-full flex gap-2 justify-between px-4 py-1 border-b border-stone-200">
-        {headers.map((header: Header, index: number) => (
-          <div key={index} className={columnClasses[header.id]}>{header.label}</div>
-        ))}
-      </div>
-
-      {rows.map((row: Row, index: number) => {
-        return <div className="w-full flex gap-2 justify-between px-4 py-1 border-b border-stone-200">
-          {headers.map((header: Header) => (
-            <div key={index} className={columnClasses[header.id]}>{row[header.id]}</div>
-          ))}
-        </div>;
-      })}
-    </div>
-  );
 }
 
 export default function KeyResults({objectiveID} : {objectiveID: string}) {
